chore(): move entry-card-removed  component
